pyproject.toml
[project]
name = "llm-personal-assistant"
version = "0.1.0"
description = ""
authors = [
    {name = "Sue Liu",email = "sue.liu@kingfisher.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "langchain (>=0.3.17,<0.4.0)",
    "jupyter (>=1.1.1,<2.0.0)",
    "langchain-community (>=0.3.16,<0.4.0)",
    "langchain-openai (>=0.3.3,<0.4.0)",
    "langchain-ollama (>=0.2.3,<0.3.0)",
    "langgraph (>=0.2.68,<0.3.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "openai (>=1.61.0,<2.0.0)",
    "youtube-transcript-api (>=0.6.3,<0.7.0)",
    "pytube (>=15.0.0,<16.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
